加实名认证

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/ysxfront/goods/GoodsCollectionController.java
@@ -5,6 +5,7 @@ import java.util.Date;
 import java.util.List;
 
 import com.ruoyi.common.base.ResponseResult;
+import com.ruoyi.common.constant.CustomerConstants;
 import com.ruoyi.common.enums.ResponseEnum;
 import com.ruoyi.web.controller.ysxfront.BaseFrontController;
 import com.ruoyi.yishengxin.Vo.GoodsCollectionVo;
@@ -50,7 +51,7 @@ public class GoodsCollectionController extends BaseFrontController {
         // 校验登录状态
         VipUser vipUser = userExist(token);
 
-        if (vipUser == null) {
+        if (null == vipUser) {
             return ResponseResult.responseResult(ResponseEnum.VIP_TOKEN_FAIL);
         }
         GoodsCollection goodsCollection = new GoodsCollection();
@@ -90,6 +91,10 @@ public class GoodsCollectionController extends BaseFrontController {
         if (null == vipUser) {
             return ResponseResult.responseResult(ResponseEnum.VIP_TOKEN_FAIL);
         }
+        if(vipUser.getIsMark().equals(CustomerConstants.NO)){
+            return ResponseResult.responseResult(ResponseEnum.IDCARD_NO_IDENTIFY);
+        }
+
         GoodsCollection goodsCollection = new GoodsCollection();
 
             GoodsCollection goodsCollection1 = new GoodsCollection();
@@ -132,9 +137,18 @@ public class GoodsCollectionController extends BaseFrontController {
         // 校验登录状态
         VipUser vipUser = userExist(token);
 
-        if (vipUser == null) {
+        if (null == vipUser) {
             return ResponseResult.responseResult(ResponseEnum.VIP_TOKEN_FAIL);
         }
+
+        if(vipUser.getIsMark().equals(CustomerConstants.NO)){
+            return ResponseResult.responseResult(ResponseEnum.IDCARD_NO_IDENTIFY);
+        }
+
+        if(vipUser.getIsMark().equals(CustomerConstants.NO)){
+            return ResponseResult.responseResult(ResponseEnum.IDCARD_NO_IDENTIFY);
+        }
+
         for (int i = 0; i < ids1.length; i++) {
             String id = ids1[i];
             int i1 = goodsCollectionService.deleteGoodsCollectionByIds(id);
@@ -167,6 +181,11 @@ public class GoodsCollectionController extends BaseFrontController {
         if (null == vipUser) {
             return ResponseResult.responseResult(ResponseEnum.VIP_TOKEN_FAIL);
         }
+
+        if(vipUser.getIsMark().equals(CustomerConstants.NO)){
+            return ResponseResult.responseResult(ResponseEnum.IDCARD_NO_IDENTIFY);
+        }
+
         Integer id = vipUser.getId();
 
         int i1 = goodsCollectionService.deleteGoodsCollectionByGid(id,gid);
@@ -198,7 +217,9 @@ public class GoodsCollectionController extends BaseFrontController {
         if (vipUser == null) {
             return ResponseResult.responseResult(ResponseEnum.VIP_TOKEN_FAIL);
         }
-
+        if(vipUser.getIsMark().equals(CustomerConstants.NO)){
+            return ResponseResult.responseResult(ResponseEnum.IDCARD_NO_IDENTIFY);
+        }
         Integer id1 = vipUser.getId();
 
         for (int i = 0; i < ids1.length; i++) {
